requirements.txt
discord.py==2.2.3
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
scikit-learn==1.3.0